refactor green tree iterators into own module
