+++ src/green/node.rs
@@ -1,13 +1,12 @@
 use std::{
     hash::{Hash, Hasher},
-    iter::FusedIterator,
     slice,
 };
 
 use fxhash::FxHasher32;
 
 use crate::{
-    green::{GreenElement, GreenElementRef, PackedGreenElement, SyntaxKind},
+    green::{iter::GreenNodeChildren, GreenElement, PackedGreenElement, SyntaxKind},
     TextSize,
 };
 use triomphe::{Arc, HeaderWithLength, ThinArc};
@@ -135,8 +134,8 @@ impl GreenNode {
 
     /// Iterator over all children of this node.
     #[inline]
-    pub fn children(&self) -> Children<'_> {
-        Children {
+    pub fn children(&self) -> GreenNodeChildren<'_> {
+        GreenNodeChildren {
             inner: self.data.slice.iter(),
         }
     }
@@ -156,90 +155,3 @@ impl PartialEq for GreenNode {
 }
 
 impl Eq for GreenNode {}
-
-/// An iterator over a [`GreenNode`]'s children.
-#[derive(Debug, Clone)]
-pub struct Children<'a> {
-    inner: slice::Iter<'a, PackedGreenElement>,
-}
-
-// NB: forward everything stable that iter::Slice specializes as of Rust 1.39.0
-impl ExactSizeIterator for Children<'_> {
-    #[inline(always)]
-    fn len(&self) -> usize {
-        self.inner.len()
-    }
-}
-
-impl<'a> Iterator for Children<'a> {
-    type Item = GreenElementRef<'a>;
-
-    #[inline]
-    fn next(&mut self) -> Option<GreenElementRef<'a>> {
-        self.inner.next().map(PackedGreenElement::as_ref)
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        self.inner.size_hint()
-    }
-
-    #[inline]
-    fn count(self) -> usize
-    where
-        Self: Sized,
-    {
-        self.inner.count()
-    }
-
-    #[inline]
-    fn nth(&mut self, n: usize) -> Option<Self::Item> {
-        self.inner.nth(n).map(PackedGreenElement::as_ref)
-    }
-
-    #[inline]
-    fn last(mut self) -> Option<Self::Item>
-    where
-        Self: Sized,
-    {
-        self.next_back()
-    }
-
-    #[inline]
-    fn fold<Acc, Fold>(self, init: Acc, mut f: Fold) -> Acc
-    where
-        Fold: FnMut(Acc, Self::Item) -> Acc,
-    {
-        let mut accum = init;
-        for x in self {
-            accum = f(accum, x);
-        }
-        accum
-    }
-}
-
-impl<'a> DoubleEndedIterator for Children<'a> {
-    #[inline]
-    fn next_back(&mut self) -> Option<Self::Item> {
-        self.inner.next_back().map(PackedGreenElement::as_ref)
-    }
-
-    #[inline]
-    fn nth_back(&mut self, n: usize) -> Option<Self::Item> {
-        self.inner.nth_back(n).map(PackedGreenElement::as_ref)
-    }
-
-    #[inline]
-    fn rfold<Acc, Fold>(mut self, init: Acc, mut f: Fold) -> Acc
-    where
-        Fold: FnMut(Acc, Self::Item) -> Acc,
-    {
-        let mut accum = init;
-        while let Some(x) = self.next_back() {
-            accum = f(accum, x);
-        }
-        accum
-    }
-}
-
-impl FusedIterator for Children<'_> {}
